+++ src/node/ext/ldap/filter.py
@@ -1,15 +1,15 @@
 # -*- coding: utf-8 -*-
+from node.ext.ldap.base import encode_utf8
+
+
 class LDAPFilter(object):
-    
-    encoding = 'utf-8'
     
     def __init__(self, queryFilter=None):
         if queryFilter is not None \
                 and not isinstance(queryFilter, basestring) \
                 and not isinstance(queryFilter, LDAPFilter):
             raise TypeError('Query filter must be LDAPFilter or string')
-        if isinstance(queryFilter, unicode):
-            queryFilter = queryFilter.encode(self.encoding)
+        queryFilter = encode_utf8(queryFilter)
         self._filter = queryFilter
         if isinstance(queryFilter, LDAPFilter):
             self._filter = str(queryFilter)
@@ -64,7 +64,7 @@ class LDAPDictFilter(LDAPFilter):
     def __str__(self):
         if not self.criteria:
             return ''
-        return str(dict_to_filter(self.criteria, self.or_search, self.encoding))
+        return str(dict_to_filter(self.criteria, self.or_search))
 
     def __repr__(self):
         return "LDAPDictFilter(criteria=%r)" % (self.criteria,)
@@ -104,29 +104,26 @@ class LDAPRelationFilter(LDAPFilter):
         if len(dictionary) is 1:
             _filter = LDAPFilter(self.relation)
         else:
-            _filter = dict_to_filter(
-                parsedRelation, self.or_search, self.encoding)
+            _filter = dict_to_filter(parsedRelation, self.or_search)
 
         return self.dictionary and \
-            str(dict_to_filter(
-                self.dictionary, self.or_search, self.encoding)) or ''
+            str(dict_to_filter(self.dictionary, self.or_search)) or ''
     
     def __repr__(self):
         return "LDAPRelationFilter('%s')" % (str(self),)
 
 
-def dict_to_filter(criteria, or_search, encoding):
+def dict_to_filter(criteria, or_search):
     """Turn dictionary criteria into ldap queryFilter string
     """
     _filter = None
     for attr, values in criteria.items():
-        if isinstance(attr, unicode):
-            attr = attr.encode(encoding)
+        attr = encode_utf8(attr)
         if not isinstance(values, list):
             values = [values]
         for value in values:
             if isinstance(value, unicode):
-                value = value.encode(encoding)
+                value = encode_utf8(value)
             if _filter is None:
                 _filter = LDAPFilter('(%s=%s)' % (attr, value))
             else: